# docs/swagger.yaml
basePath: /
definitions:
  advert_model.Advert:
    properties:
      body:
        type: string
      id:
        type: string
      image:
        type: string
      price:
        type: integer
      title:
        type: string
    type: object
  advert_model.AdvertInput:
    properties:
      body:
        type: string
      image:
        type: string
      price:
        type: integer
      title:
        type: string
    type: object
  advert_model.AdvertWithOwner:
    properties:
      body:
        type: string
      id:
        type: string
      image:
        type: string
      own:
        type: boolean
      price:
        type: integer
      title:
        type: string
      user_id:
        type: string
      user_login:
        type: string
    type: object
  transport.RespWriter:
    properties:
      message:
        type: string
    type: object
  transport.SignUpResponse:
    properties:
      id:
        type: string
      login:
        type: string
    type: object
  user_model.UserLogin:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API server for Маркетплейс
  title: Маркетплейс
paths:
  /advert/add:
    post:
      consumes:
      - application/json
      description: Добавление объявления в систему. Добавлять могут только авторизованные
        пользователи.
      operationId: add-advert
      parameters:
      - description: поля объявления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/advert_model.AdvertInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/advert_model.Advert'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.RespWriter'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/transport.RespWriter'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.RespWriter'
      security:
      - ApiKeyAuth: []
      summary: AddAdvert
      tags:
      - advert
  /advert/get:
    get:
      consumes:
      - application/json
      description: |-
        Получение списка объявлений. Возможность сортировки по дате и цене, также их направление.
        Возможность фильтрации по цене с мин и макс значениями.
        Сортировка задается параметрами URL: order и dir.
        Если не задать эти параметры, то будет сортировка по убыванию времени
      operationId: get-adverts
      parameters:
      - description: type of order
        in: query
        name: order
        type: string
      - description: asc or desc
        in: query
        name: dir
        type: string
      - description: min price
        in: query
        name: min
        type: string
      - description: max price
        in: query
        name: max
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/advert_model.AdvertWithOwner'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.RespWriter'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.RespWriter'
      security:
      - ApiKeyAuth: []
      summary: getAdverts
      tags:
      - advert
  /login:
    post:
      consumes:
      - application/json
      description: Вход в систему. Указывается логин и пароль
      operationId: sign-in
      parameters:
      - description: информация о пользователе
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user_model.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.RespWriter'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.RespWriter'
      summary: SignIn
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: |-
        Регистрациия в системе. Минимальная длина логина и пароля - 3 символа.
        Максимальная длина - 50 символов. Логин должен быть уникальным.
      operationId: sign-up
      parameters:
      - description: информация о пользователе
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/user_model.UserLogin'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/transport.SignUpResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/transport.RespWriter'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/transport.RespWriter'
      summary: SignUp
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"